Moved most of Interop.Win32 into Rylogic.Core
Rylogic.View3d now doesn't depend on Rylogic.Core.Windows
Changed to using doubles for Frac/Lerp

+++ projects/Rylogic.Core/src/Maths/Matrix3x4.cs
@@ -529,7 +529,7 @@ namespace Rylogic.Maths
 		}
 
 		/// <summary>Spherically interpolate between two rotations</summary>
-		public static m3x4 Slerp(m3x4 lhs, m3x4 rhs, float frac)
+		public static m3x4 Slerp(m3x4 lhs, m3x4 rhs, double frac)
 		{
 			return new m3x4(Slerp(new quat(lhs), new quat(rhs), frac));
 		}
